requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
python-dotenv==1.0.1
sentence-transformers>=3.0.0
PyPDF2==3.0.1
python-docx==1.1.0
qdrant-client>=1.7.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
